Refactor marker data access
