+++ tests/op_datasets/etl/blockbatch/test_construct_tasks_blockbatch.py
@@ -32,7 +32,7 @@ def make_dataframe(path: str):
 
 
 def test_construct():
-    with patch("op_analytics.coreutils.partitioned.dataaccess.run_query_oplabs") as m1:
+    with patch("op_analytics.coreutils.partitioned.markers_clickhouse.run_query_oplabs") as m1:
         m1.return_value = make_dataframe("mainnet_markers.json")
         tasks = construct_tasks(
             chains=["mode"],
